Add a unique name to objects so we can determine uniqueness when read back in
    In Nagios it is possible to have multiple service checks with the
same service_description. eg I could have an check with a
service_description of 'SSH' for multiple hosts. So in puppet we can't
use it as a unique name for the resource. This patch modifies the code
to use $name as the unique name. For some types eg command_name $name
ends up in the config and thus we can tell which puppet resources match
to which nagios ones.  For other types like service there is no direct
mapping from $name to a nagios attibute. So we use a custom attribute
called _naginator_name.

Signed-off-by: John Ferlito <johnf@inodes.org>